Simplified Code, Added Remove Card functionality.
Simplified a lot of the code and made it look nicer. Mostly added the functionality to remove the cards and made it so the numbers have unique keys as well as so they aren't so random like they where before. Now it'll keep each generated number through each render instead of always rendering a new set of random numbers for each card.